+++ holoviews/ipython/preprocessors.py
@@ -1,12 +1,11 @@
 """
 Prototype demo:
 
-python convert.py Conversion_Example.ipynb > foo.py ; ipython foo.py ; open foo.png
+python holoviews/ipython/convert.py Conversion_Example.ipynb | python
 """
 import nbconvert, nbformat
 from nbconvert.preprocessors import Preprocessor
-import sys
-import ast
+import os, sys, ast
 
 
 
@@ -51,7 +50,8 @@ def wrap_cell_expression(source, template='{expr}'):
                               + ([expr_end_slice] if expr_end_slice else []))
             ending = '\n'.join(([expr_start_slice] if expr_start_slice else [])
                             + filtered[last_expr.lineno:])
-            return start + template.format(expr=ending)
+            # BUG!! Adds newline for 'foo'; <expr>
+            return start + '\n' + template.format(expr=ending)
     return source
 
 
@@ -146,8 +146,59 @@ class StripMagicsProcessor(Preprocessor):
     def __call__(self, nb, resources): return self.preprocess(nb,resources)
 
 
+class Substitute(Preprocessor):
+    """
+    An nbconvert preprocessor that substitutes one set of HTML data
+    output for another, adding annotation to the output as required.
+
+    The constructor accepts the notebook format version and a
+    substitutions dictionary:
+
+    {source_html:(target_html, annotation)}
+
+    Where the annotation may be None (i.e. no annotation).
+    """
+    annotation = '<center><b>%s</b></center>'
+
+    def __init__(self, version, substitutions, **kw):
+        self.nbversion = version
+        self.substitutions = substitutions
+        super(Preprocessor, self).__init__(**kw)
+
+    def __call__(self, nb, resources): # Temporary hack around 'enabled' flag
+        return self.preprocess(nb,resources)
+
+
+    def replace(self, src):
+        "Given some source html substitute and annotated as applicable"
+        for html in self.substitutions.keys():
+            if src == html:
+                annotation = self.annotation % self.substitutions[src][1]
+                return annotation + self.substitutions[src][0]
+        return src
+
+
+    def preprocess_cell(self, cell, resources, index):
+        v4 = (self.nbversion[0] == 4)
+        if cell['cell_type'] == 'code':
+            for outputs in cell['outputs']:
+                output_key = ('execute_result' if v4 else 'pyout')
+                if outputs['output_type'] == output_key:
+                    # V1-3
+                    if not v4 and 'html' in outputs:
+                        outputs['html'] = self.replace(outputs['html'])
+                    # V4
+                    for data in outputs.get('data',[]):
+                        if v4 and data == 'text/html':
+                            substitution = self.replace(outputs['data']['text/html'])
+                            outputs['data']['text/html'] = substitution
+        return cell, resources
+
+
 if __name__ == '__main__':
-    with open(sys.argv[1]) as f:
+    filename = sys.argv[1]
+    basename = os.path.splitext(os.path.basename(filename))[0]
+    with open(filename) as f:
         nb = nbformat.read(f, nbformat.NO_CONVERT)
         exporter = nbconvert.PythonExporter()
         preprocessors = [OptsMagicProcessor(),
